openbench/services/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
SQLAlchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
aiosqlite==0.20.0
psycopg2-binary==2.9.10
greenlet==3.1.1
pgvector==0.3.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
email-validator==2.2.0
httpx==0.28.1
celery[redis]==5.4.0
redis==5.2.1
boto3==1.35.92
botocore==1.35.92
anthropic==0.43.1
pdfplumber==0.11.5
PyMuPDF==1.25.1
python-docx==1.1.2
pillow==11.0.0
pytesseract==0.3.13
markdown==3.7
bleach==6.2.0
markdown-it-py==3.0.0
weasyprint==63.1
jinja2==3.1.5
tenacity==9.0.0
loguru==0.7.3
orjson==3.10.13
python-slugify==8.0.4
pytest==8.3.4
pytest-asyncio==0.25.0
ulid-py==1.1.0
itsdangerous==2.2.0
PyJWT==2.10.1
rapidfuzz==3.10.1
